# openapi.yml
openapi: 3.1.0
info:
  title: CubeAPI
  description: OpenAPI contract for the CubeSandbox dashboard surface.
  license:
    name: ''
  version: 0.1.0
servers:
- url: /cubeapi/v1
  description: CubeAPI dashboard surface
paths:
  /cluster/overview:
    get:
      tags:
      - handlers::cluster
      operationId: cluster_overview
      responses:
        '200':
          description: Cluster capacity overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterOverview'
        '404':
          description: Cluster endpoint unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /nodes:
    get:
      tags:
      - handlers::cluster
      operationId: list_nodes
      responses:
        '200':
          description: Node list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NodeView'
        '404':
          description: Node inventory unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /sandboxes/{sandboxID}:
    get:
      tags:
      - handlers::sandboxes
      operationId: get_sandbox
      parameters:
      - name: sandboxID
        in: path
        description: Sandbox identifier
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Sandbox detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SandboxDetail'
        '404':
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - handlers::sandboxes
      operationId: kill_sandbox
      parameters:
      - name: sandboxID
        in: path
        description: Sandbox identifier
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Sandbox deleted
        '404':
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '408':
          description: The existing standard API request timeout expired before the synchronous delete completed
        '409':
          description: Paused sandbox cannot be admitted for internal resume because node capacity or resource metadata is unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: Sandbox is pausing or another lifecycle operation is in progress (Retry-After 2); the Cubelet RPC has too little remaining time or its internal resume could not be completed (Retry-After 5)
          headers:
            Retry-After:
              schema:
                type: integer
                format: int64
                minimum: 0
              description: Seconds a client should wait before retrying DELETE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /sandboxes/{sandboxID}/pause:
    post:
      tags:
      - handlers::sandboxes
      operationId: pause_sandbox
      parameters:
      - name: sandboxID
        in: path
        description: Sandbox identifier
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Sandbox paused
        '404':
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Sandbox cannot be paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /sandboxes/{sandboxID}/resume:
    post:
      tags:
      - handlers::sandboxes
      operationId: resume_sandbox
      parameters:
      - name: sandboxID
        in: path
        description: Sandbox identifier
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumedSandbox'
        required: true
      responses:
        '201':
          description: Sandbox resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
        '404':
          description: Sandbox not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Sandbox is already running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /templates:
    get:
      tags:
      - handlers::templates
      operationId: list_templates
      parameters:
      - name: instance_type
        in: query
        description: |-
          Optional CubeMaster instance_type filter (currently no server-side filter;
          reserved for future use).
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Template list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemplateSummary'
        '404':
          description: Template endpoint unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - handlers::templates
      operationId: create_template
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
        required: true
      responses:
        '202':
          description: Template build job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateBuildJob'
        '400':
          description: Invalid request (bad alias, missing image, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /templates/aliases/{alias}:
    get:
      tags:
      - handlers::templates
      operationId: get_template_by_alias
      parameters:
      - name: alias
        in: path
        description: Template alias
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Template alias lookup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateAliasLookupResponse'
        '400':
          description: Invalid alias
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Template alias not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /templates/compat:
    get:
      tags:
      - handlers::templates
      operationId: template_compat
      responses:
        '200':
          description: Template compatibility matrix
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateCompatMatrixView'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /templates/compat/{templateID}/adopt-baseline:
    post:
      tags:
      - handlers::templates
      operationId: adopt_template_compat_baseline
      parameters:
      - name: templateID
        in: path
        description: Template identifier
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Adopted UNKNOWN replicas to current baseline
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateCompatAdoptResponseView'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /templates/{templateID}:
    get:
      tags:
      - handlers::templates
      operationId: get_template
      parameters:
      - name: templateID
        in: path
        description: Template identifier
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Template detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateDetail'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - handlers::templates
      operationId: rebuild_template
      parameters:
      - name: templateID
        in: path
        description: Template identifier
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RebuildTemplateRequest'
        required: true
      responses:
        '202':
          description: Rebuild job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateBuildJob'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - handlers::templates
      operationId: delete_template
      parameters:
      - name: templateID
        in: path
        description: Template identifier or alias
        required: true
        schema:
          type: string
      - name: instance_type
        in: query
        description: CubeMaster instance_type filter
        required: false
        schema:
          type: string
      - name: sync
        in: query
        description: Wait for deletion to complete before returning
        required: false
        schema:
          type: boolean
      responses:
        '204':
          description: Template deleted
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - handlers::templates
      operationId: update_template
      parameters:
      - name: templateID
        in: path
        description: Template identifier
        required: true
        schema:
          type: string
      responses:
        '501':
          description: Not implemented; use POST /templates/{id} to rebuild
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /templates/{templateID}/builds/{buildID}:
    post:
      tags:
      - handlers::templates
      operationId: start_template_build
      parameters:
      - name: templateID
        in: path
        description: Template identifier
        required: true
        schema:
          type: string
      - name: buildID
        in: path
        description: Build identifier
        required: true
        schema:
          type: string
      responses:
        '202':
          description: Build started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateBuildJob'
        '404':
          description: Template or build not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /templates/{templateID}/builds/{buildID}/logs:
    get:
      tags:
      - handlers::templates
      operationId: get_template_build_logs
      parameters:
      - name: templateID
        in: path
        description: Template identifier
        required: true
        schema:
          type: string
      - name: buildID
        in: path
        description: Build identifier
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Build logs (JSON)
        '404':
          description: Build not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    ClusterOverview:
      type: object
      required:
      - nodeCount
      - healthyNodes
      - totalCpuMilli
      - allocatableCpuMilli
      - totalMemoryMB
      - allocatableMemoryMB
      - maxMvmSlots
      properties:
        allocatableCpuMilli:
          type: integer
          format: int64
          description: Currently-allocatable CPU in millicpu.
        allocatableMemoryMB:
          type: integer
          format: int64
          description: Currently-allocatable memory in MiB.
        healthyNodes:
          type: integer
          minimum: 0
        maxMvmSlots:
          type: integer
          format: int64
          description: Sum of every node's maximum MVM slots.
        nodeCount:
          type: integer
          minimum: 0
        totalCpuMilli:
          type: integer
          format: int64
          description: Total CPU across the cluster, expressed in millicpu.
        totalMemoryMB:
          type: integer
          format: int64
          description: Total memory in MiB.
...
    RebuildTemplateRequest:
      type: object
      description: Body for POST /templates/:id (rebuild).
      additionalProperties: {}
    ResumedSandbox:
      type: object
      description: Request body for POST /sandboxes/{id}/resume (deprecated).
      properties:
        autoPause:
          type: boolean
        timeout:
          type: integer
          format: int32
...
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...